# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "EduCRM",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://edu-crm-api-exp.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "GroupClass": {
        "type": "object",
        "properties": {
          "groupCourse": {
            "type": "string",
            "description": "The ID of the group course associated with the group class"
          },
          "status": {
            "type": "string",
            "enum": ["not_started", "ongoing", "finished"],
            "description": "The status of the group class"
          },
          "starting_date": {
            "type": "string",
            "format": "date",
            "description": "The starting date of the group class"
          },
          "finish_date": {
            "type": "string",
            "format": "date",
            "description": "The finish date of the group class"
          },
          "sessions": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "The IDs of the group class sessions associated with the group class"
            },
            "description": "The sessions of the group class"
          },
          "students": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "The IDs of the student profiles enrolled in the group class"
            },
            "description": "The students enrolled in the group class"
          },
          "intake_gallery": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Links to images representing the intake gallery of the group class"
            },
            "description": "The intake gallery of the group class"
          },
          "testimonials": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "The IDs of the student testimonials associated with the group class"
            },
            "description": "The testimonials of the group class"
          },
          "price": {
            "type": "string",
            "description": "The price of the group class"
          }
        }
      },
...
      "GroupClassSession": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "default": 0
          },
          "objective": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "attendance": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "student": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": ["absent", "present", "excused"]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin": {
      "post": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "201": {
            "description": "Curator registered"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Can't store user in database"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Curator deleted"
          },
          "500": {
            "description": "Can't delete `curator` from database"
          }
        }
      }
    },
    "/admin/academicArea": {
      "post": {
        "tags": [
          "Academic Area"
        ],
        "responses": {
          "201": {
            "description": "Academic area was added"
          },
          "500": {
            "description": "Can't store `Academic Area` in database"
          }
        }
      },
      "delete": {
        "tags": [
          "Academic Area"
        ],
        "responses": {
          "200": {
            "description": "Academic area deleted"
          },
          "500": {
            "description": "Can't delete `academic area`"
          }
        }
      }
    },
...
    "/admin/classRequest": {
      "put": {
        "tags": [
          "OneToOne"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string",
                    "description": "The ID of the request to be updated."
                  },
                  "status": {
                    "type": "string",
                    "description": "The new status to be assigned to the request. Must be either 'approved' or 'declined'."
                  }
                },
                "required": [
                  "requestId",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status of student's request changed"
          },
          "404": {
            "description": "Incorrect request ID"
          },
          "500": {
            "description": "Can't update request"
          }
        }
      }
    },
...
    "/admin/students": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "View all students",
        "responses": {
          "200": {
            "description": "Successfully got student profiles"
          },
          "404": {
            "description": "No Data"
          }
        }
      }
    },
...
    "/admin/contactServices": {
      "post": {
        "tags": [
          "Util"
        ],
        "summary": "create a new contact service. i.e., Instagram, Telegram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contactService": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Database error, while saving `ContactService`"
          }
        }
      }
    },
    "/admin/contactServices/{id}": {
      "put": {
        "tags": [
          "Util"
        ],
        "summary": "edit existing contact service. i.e., fix a typo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the contact service to be edited"
          },
          {
            "in": "body",
            "name": "contactService",
            "required": true,
            "schema": {
              "type": "object"
            },
            "description": "The updated contact service data"
          }
        ],
        "responses": {
          "200": {
            "description": "ContactService was updated"
          },
          "400": {
            "description": "ContactService doesn't exist"
          }
        }
      },
      "delete": {
        "tags": [
          "Util"
        ],
        "summary": "delete a contact service by service id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ContactService was deleted"
          },
          "500": {
            "description": "Database error deleting `ContactService`"
          }
        }
      }
    },
...
    "/admin/payment": {
      "post": {
        "tags": [
          "OneToOne"
        ],
        "summary": "Request payment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "student": {
                    "type": "string",
                    "description": "The ID of the student making the payment request."
                  },
                  "request": {
                    "type": "string",
                    "description": "The ID of the one-to-one request for which payment is being requested."
                  }
                },
                "required": [
                  "student",
                  "request"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment request saved"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Request is not approved"
          },
          "404": {
            "description": "Class request lost"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/admin/group_class/{id}": {
      "delete": {
        "tags": [
          "Group Course"
        ],
        "summary": "Delete a group class. Access: admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the group class to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Group class successfully deleted"
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "put": {
        "tags": ["Group Course"],
        "summary": "Update a group class by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group class to update"
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupClass"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GroupClass updated successfully"
          },
          "400": {
            "description": "GroupClass not found"
          }
        }
      },
      "get": {
        "tags": ["Group Course"],
        "summary": "Get group class by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group class to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Retrieved group class."
          },
          "404": {
            "description": "Group class not found."
          }
        }
      }
    },
    "/admin/group_session/{id}": {
      "delete": {
        "tags": [
          "Group Course"
        ],
        "summary": "Delete a group session.Access: admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the group session to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Group session successfully deleted"
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "put": {
        "tags": ["Group Course"],
        "summary": "Update an existing group class session",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization token"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group class session to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupClassSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group class session updated successfully"
          },
          "400": {
            "description": "GroupClassSession doesn't exist"
          }
        }
      },
      "get": {
        "tags": ["Group Course"],
        "summary": "Get group session by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group session to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Retrieved group session."
          },
          "403": {
            "description": "You are not authorized to access this group session."
          },
          "404": {
            "description": "Group session not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  "/admin/group_class": {
    "post": {
      "tags": [
        "Group Class"
      ],
      "summary": "Create a new group class.Access: admin",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GroupClass"
            }
          }
        },
        "required": true
      },
      "responses": {
        "201": {
          "description": "Group class created successfully"
        },
        "500": {
          "description": "Database error"
        }
      }
    },
    "get": {
      "tags": ["Group Course"],
      "summary": "Get all group classes",
      "responses": {
        "200": {
          "description": "Success. Retrieved group classes."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
...
    "/admin/group_session": {
      "post": {
        "tags": ["Group Course"],
        "summary": "Create a new group class session",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupClassSession"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group class session created successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": ["Group Course"],
        "summary": "Get all group sessions",
        "responses": {
          "200": {
            "description": "Success. Retrieved group sessions."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/student/group_class": {
      "get": {
        "tags": ["Group Course"],
        "summary": "Get all group classes",
        "responses": {
          "200": {
            "description": "Success. Retrieved group classes."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/student/group_session": {
      "get": {
        "tags": ["Group Course"],
        "summary": "Get all group sessions",
        "responses": {
          "200": {
            "description": "Success. Retrieved group sessions."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/student/group_session/{id}": {
      "get": {
        "tags": ["Group Course"],
        "summary": "Get group session by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group session to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Retrieved group session."
          },
          "403": {
            "description": "You are not authorized to access this group session."
          },
          "404": {
            "description": "Group session not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
},
    "/student/group_class/{id}": {
      "get": {
        "tags": ["Group Course"],
        "summary": "Get group class by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the group class to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Retrieved group class."
          },
          "404": {
            "description": "Group class not found."
          }
        }
      }
    },
...
  }
}